# Unroll secp256k1_fe_(get|set)_b32 to make them much faster.