@sipa 2017-01-18
I believe that the benchmark results are due to the entire loop being optimized out (reducing the iteration count increases the timings proportionally).