@sipa 2017-02-17
Can you adapt the benchmark so it's not optimized away?