@gmaxwell 2017-04-24
I stripped the benchmarks out.