@sipa 2017-04-25
Benchmarked `bench_verify` with GMP disabled on a i7-6820HQ CPU, pegged to 2.6GHz.
* master: 97.7us (64-bit), 366us (32-bit).
* this PR: 97.5us (64-bit), 365us (32-bit).
So a 0.2% (64-bit) and 0.3% improvement (32-bit).